+++ internal/pkg/instrumentation/bpf/github.com/gin-gonic/gin/bpf/probe.bpf.c
@@ -72,11 +72,22 @@ int uprobe_GinEngine_ServeHTTP(struct pt_regs *ctx) {
         return 0;
     }
 
+    // Propagate context
+    struct span_context *parent_ctx = get_parent_span_context(req_ctx_ptr);
+    if (parent_ctx != NULL)
+    {
+        // found parent context
+        httpReq.psc = *parent_ctx;
+        copy_byte_arrays(httpReq.psc.TraceID, httpReq.sc.TraceID, TRACE_ID_SIZE);
+        generate_random_bytes(httpReq.sc.SpanID, SPAN_ID_SIZE);
+    }
+    else
+    {
+        httpReq.sc = generate_span_context();
+    }
+
     // Get key
     void *key = get_consistent_key(ctx, req_ctx_ptr);
-
-    // Write event
-    httpReq.sc = generate_span_context();
     bpf_map_update_elem(&http_events, &key, &httpReq, 0);
     start_tracking_span(req_ctx_ptr, &httpReq.sc);
     return 0;