+++ internal/pkg/instrumentation/bpf/github.com/gin-gonic/gin/probe_test.go
@@ -18,9 +18,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/go-logr/logr/testr"
 	"github.com/stretchr/testify/assert"
-
 	"go.opentelemetry.io/otel/attribute"
 	semconv "go.opentelemetry.io/otel/semconv/v1.21.0"
 	"go.opentelemetry.io/otel/trace"
@@ -36,8 +34,7 @@ func TestProbeConvertEvent(t *testing.T) {
 	traceID := trace.TraceID{1}
 	spanID := trace.SpanID{1}
 
-	i := New(testr.New(t))
-	got := i.convertEvent(&Event{
+	got := convertEvent(&event{
 		BaseSpanProperties: context.BaseSpanProperties{
 			StartTime:   uint64(start.UnixNano()),
 			EndTime:     uint64(end.UnixNano()),
@@ -55,7 +52,7 @@ func TestProbeConvertEvent(t *testing.T) {
 		TraceFlags: trace.FlagsSampled,
 	})
 	want := &probe.Event{
-		Library:     instrumentedPkg,
+		Package:     pkg,
 		Name:        "GET",
 		Kind:        trace.SpanKindServer,
 		StartTime:   int64(start.UnixNano()),