+++ internal/test/test_helpers/utilities.bash
@@ -6,11 +6,39 @@ span_names_for() {
 	spans_from_scope_named $1 | jq '.name'
 }
 
+# Returns a list of server span names emitted by a given library/scope
+	# $1 - library/scope name
+server_span_names_for() {
+	server_spans_from_scope_named $1 | jq '.name'
+}
+
+# Returns a list of client span names emitted by a given library/scope
+	# $1 - library/scope name
+client_span_names_for() {
+	client_spans_from_scope_named $1 | jq '.name'
+}
+
 # Returns a list of attributes emitted by a given library/scope
 span_attributes_for() {
 	# $1 - library/scope name
 
 	spans_from_scope_named $1 | \
+		jq ".attributes[]"
+}
+
+# Returns a list of attributes emitted by a given library/scope on server spans.
+server_span_attributes_for() {
+	# $1 - library/scope name
+
+	server_spans_from_scope_named $1 | \
+		jq ".attributes[]"
+}
+
+# Returns a list of attributes emitted by a given library/scope on clinet_spans.
+client_span_attributes_for() {
+	# $1 - library/scope name
+
+	client_spans_from_scope_named $1 | \
 		jq ".attributes[]"
 }
 
@@ -23,6 +51,18 @@ resource_attributes_received() {
 	# $1 - library/scope name
 spans_from_scope_named() {
 	spans_received | jq ".scopeSpans[] | select(.scope.name == \"$1\").spans[]"
+}
+
+# Returns an array of all server spans emitted by a given library/scope
+	# $1 - library/scope name
+server_spans_from_scope_named() {
+	spans_from_scope_named $1 | jq "select(.kind == 2)"
+}
+
+# Returns an array of all client spans emitted by a given library/scope
+	# $1 - library/scope name
+client_spans_from_scope_named() {
+	spans_from_scope_named $1 | jq "select(.kind == 3)"
 }
 
 # Returns an array of all spans received
@@ -42,8 +82,7 @@ redact_json() {
 		jq --sort-keys '
 			del(
 				.resourceSpans[].scopeSpans[].spans[].startTimeUnixNano,
-				.resourceSpans[].scopeSpans[].spans[].endTimeUnixNano,
-				(.resourceSpans[].scopeSpans[].spans[].attributes[] | select(.key == "net.peer.port"))
+				.resourceSpans[].scopeSpans[].spans[].endTimeUnixNano
 			)
 			| .resourceSpans[].scopeSpans[].spans[].traceId|= (if
 					. // "" | test("^[A-Fa-f0-9]{32}$") then "xxxxx" else (. + "<-INVALID")
@@ -54,7 +93,13 @@ redact_json() {
 			| .resourceSpans[].scopeSpans[].spans[].parentSpanId|= (if
 					. // "" | test("^[A-Fa-f0-9]{16}$") then "xxxxx" else (. + "")
 				end)
+			| .resourceSpans[].scopeSpans[].spans[].attributes[] |= if 
+					(.key == "net.peer.port") then .value.intValue |= (if
+				   		. // "" | test("^[1-9][0-9]{0,4}$") then "xxxxx" else (. + "") 
+					end) else . 
+				end
 			| .resourceSpans[].scopeSpans|=sort_by(.scope.name)
+			| .resourceSpans[].scopeSpans[].spans|=sort_by(.kind)
 			' > ${BATS_TEST_DIRNAME}/traces.json
 }
 
@@ -65,6 +110,9 @@ MATCH_A_TRACE_ID=^"\"[A-Fa-f0-9]{32}\"$"
 
 # expect a 16-digit hexadecimal string (in quotes)
 MATCH_A_SPAN_ID=^"\"[A-Fa-f0-9]{16}\"$"
+
+# ecpect a valid port number
+MATCH_A_PORT_NUMBER=^"\"[1-9][0-9]{0,4}\"$"
 
 # Fail and display details if the expected and actual values do not
 # equal. Details include both values.